+++ packages/xchain-client/src/protocols.ts
@@ -1,0 +1,3 @@
+export enum Protocol {
+  THORCHAIN = 1,
+}