+++ packages/xchain-utxo/src/types/index.ts
@@ -1,0 +1,3 @@
+import { PreparedTx, UTXO, UtxoClientParams, Witness } from './types'
+
+export { UTXO, UtxoClientParams, Witness, PreparedTx }