+++ packages/xchain-utxo/src/types/types.ts
@@ -1,0 +1,13 @@
+import { ExplorerProviders, PreparedTx as BasePreparedTx, XChainClientParams } from '@xchainjs/xchain-client'
+import { UTXO, UtxoOnlineDataProviders, Witness } from '@xchainjs/xchain-utxo-providers'
+
+export type UtxoClientParams = XChainClientParams & {
+  explorerProviders: ExplorerProviders
+  dataProviders: UtxoOnlineDataProviders[]
+}
+
+export type PreparedTx = BasePreparedTx & {
+  utxos: UTXO[]
+}
+
+export { UTXO, Witness }