+++ packages/xchain-binance/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v5.7.7 (2023-12-11)
+
+## Update
+
+- Client dependency updated
+
 # v5.7.6 (2023-12-01)
 
 ## Update