+++ packages/xchain-bitcoin/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.23.8 (2023-12-11)
+
+## Update
+
+- UTXO client package dependency
+
 # v0.23.7 (2023-12-06)
 
 ## Add