+++ packages/xchain-dash/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.2.7 (2023-12-11)
+
+## Update
+
+- UTXO client package dependency
+
 # v0.2.6 (2023-12-06)
 
 ## Update