+++ packages/xchain-doge/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.7.8 (2023-12-11)
+
+## Update
+
+- UTXO client package dependency
+
 # v0.7.7 (2023-11-21)
 
 ## Update