+++ packages/xchain-thorchain/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.28.15 (2023-12-11)
+
+## Update
+
+- Client and Cosmos client dependency update
+
 # v0.28.14 (2023-12-07)
 
 ## Update